// src/types/group.ts
export type NotificationPolicy = 'all_group_messages' | 'mentioned_members' | 'direct_messages_only' | 'disabled';

export type ChatGroup = {
    id: string;
    name: string;
    photoUrl: string;
    ownerId: string;
    memberIds: string[];
    memberLimit: number;
    notificationPolicy: NotificationPolicy;
    createdAt: number;
    updatedAt: number;
};